+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-06-19 23:04:31
-// Commit: chore(test): update dependencies and cleanup unused imports in test
+// Updated at: 2026-06-21 10:55:45
+// Commit: feat(test): integrate custom telemetry logger in test
 
 class testComponentState {
-  final String commitRef = '2164';
-  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
-  final DateTime timestamp = DateTime.parse('2026-06-19T23:04:31+05:00');
+  final String commitRef = '2169';
+  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
+  final DateTime timestamp = DateTime.parse('2026-06-21T10:55:45+05:00');
 }